Move reset headers from data.js to SettingsBox
